# docs/swagger.yaml
definitions:
  handlers.HealthResponse:
    properties:
      service:
        type: string
      status:
        type: string
      timestamp:
        type: string
      version:
        type: string
    type: object
  handlers.createUserRequest:
    properties:
      name:
        type: string
    type: object
  handlers.depositRequest:
    properties:
      amount:
        type: number
    type: object
  handlers.transferRequest:
    properties:
      amount:
        type: number
      description:
        type: string
      to_wallet_id:
        type: string
    type: object
  handlers.withdrawRequest:
    properties:
      amount:
        type: number
    type: object
  models.Transaction:
    properties:
      amount:
        type: number
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      reference_id:
        type: string
      type:
        description: deposit, withdraw, transfer_in, transfer_out
        type: string
      wallet_id:
        type: string
    type: object
  models.UserWithWallet:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      wallet:
        $ref: '#/definitions/models.Wallet'
    type: object
  models.Wallet:
    properties:
      balance:
        type: number
      created_at:
        type: string
      id:
        type: string
      user_id:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/users:
    post:
      consumes:
      - application/json
      parameters:
      - description: User details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.createUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserWithWallet'
      summary: Create user
      tags:
      - users
  /api/v1/wallets/{id}/balance:
    get:
      parameters:
      - description: Wallet ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Wallet'
      summary: Get wallet balance
      tags:
      - wallets
  /api/v1/wallets/{id}/deposit:
    post:
      consumes:
      - application/json
      parameters:
      - description: Wallet ID
        in: path
        name: id
        required: true
        type: string
      - description: Deposit details
        in: body
        name: deposit
        required: true
        schema:
          $ref: '#/definitions/handlers.depositRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Wallet'
      summary: Deposit to wallet
      tags:
      - wallets
  /api/v1/wallets/{id}/transactions:
    get:
      parameters:
      - description: Wallet ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Transaction'
            type: array
      summary: Get wallet transaction history
      tags:
      - wallets
  /api/v1/wallets/{id}/transfer:
    post:
      consumes:
      - application/json
      parameters:
      - description: Wallet ID
        in: path
        name: id
        required: true
        type: string
      - description: Transfer details
        in: body
        name: transfer
        required: true
        schema:
          $ref: '#/definitions/handlers.transferRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Wallet'
      summary: Transfer between wallets
      tags:
      - wallets
  /api/v1/wallets/{id}/withdraw:
    post:
      consumes:
      - application/json
      parameters:
      - description: Wallet ID
        in: path
        name: id
        required: true
        type: string
      - description: Withdraw details
        in: body
        name: withdraw
        required: true
        schema:
          $ref: '#/definitions/handlers.withdrawRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Wallet'
      summary: Withdraw from wallet
      tags:
      - wallets
  /health:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.HealthResponse'
      summary: Health check
      tags:
      - health
swagger: "2.0"